+++ plt/plt-scheduler/tests/migration_p10_to_p11.rs
@@ -9,6 +9,7 @@ use concordium_base::protocol_level_tokens::{
     CborHolderAccount, TokenAmount, TokenAuthorizations, TokenId, TokenOperation,
     TokenSupplyUpdateDetails,
 };
+use plt_block_state::entity::block_state::migration;
 use plt_block_state::entity::block_state::p10::BlockStateP10;
 use plt_block_state::entity::entity_test_stub;
 use plt_scheduler::TOKEN_MODULE_REF;
@@ -38,7 +39,8 @@ fn test_migrate_p10_to_p11() {
     );
 
     // Migrate the block state from P10 to P11.
-    let (mut context, mut block_state) = utils::migrate_p10_to_p11(&mut context, block_state);
+    let (mut context, mut block_state) =
+        migration::test_utils::migrate_p10_to_p11(&mut context, block_state);
 
     // Query the migrated state (P11) using the scheduler-implemented queries.
     assert_eq!(block_state.query_plt_list(&context), vec![token_id.clone()]);
